+++ datis/src/lib.rs
@@ -6,6 +6,8 @@ extern crate log;
 #[macro_use]
 extern crate const_cstr;
 #[macro_use]
+extern crate serde_json;
+#[macro_use]
 extern crate serde_derive;
 
 #[macro_use]